backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
python-multipart
pydantic
pydantic-settings
scikit-learn
numpy
cryptography
email-validator
slowapi
aiosqlite
bcrypt
